refactor: move errors to utils and add GroupNotFound
