+++ pkg/utils/errors.go
@@ -1,6 +1,7 @@
-package inmemory
+package utils
 
 import "errors"
 
+var GroupNotFoundError = errors.New("Group not found")
 var KeyNotFoundError = errors.New("key not found")
 var NoKeysError = errors.New("no keys found")
